# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gaming Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/payment-settings/{method}:
    put:
      operationId: updatePaymentSetting
      tags: [payment-settings, admin]
      summary: Update payment method settings (admin)
      parameters:
        - name: method
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentSettingBody"
      responses:
        "200":
          description: Updated payment setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSetting"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin or current user)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected, completed]
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /payments/upload:
    post:
      operationId: uploadPaymentScreenshot
      tags: [payments]
      summary: Upload payment screenshot for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadPaymentBody"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify or reject a payment (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Coupons ───────────────────────────────────────────────────────────────
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: List of coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    put:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResult"

  # ── Banners ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
          enum: [game-topups, gift-cards, subscriptions, vouchers]
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        dynamicFields:
          type: array
          items:
            $ref: "#/components/schemas/DynamicField"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        inStock:
          type: boolean
        featured:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, price, category, tags, dynamicFields, variants, inStock, featured, createdAt]
    DynamicField:
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, email]
        required:
          type: boolean
      required: [name, label, type, required]
    ProductVariant:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        badge:
          type: string
          nullable: true
        originalPrice:
          type: number
          nullable: true
        icon:
          type: string
          nullable: true
      required: [name, price]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
          nullable: true
        guestName:
          type: string
          nullable: true
        guestEmail:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        gameDetails:
          type: object
          additionalProperties:
            type: string
        totalAmount:
          type: number
        discountAmount:
          type: number
        couponCode:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, verified, rejected, completed]
        paymentScreenshotUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, productId, product, gameDetails, totalAmount, discountAmount, status, createdAt, updatedAt]
    CreateOrderBody:
      type: object
      properties:
        productId:
          type: integer
        guestName:
          type: string
          nullable: true
        guestEmail:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        gameDetails:
          type: object
          additionalProperties:
            type: string
        couponCode:
          type: string
          nullable: true
      required: [productId, gameDetails]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, verified, rejected, completed]
        adminNote:
          type: string
          nullable: true
      required: [status]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        screenshotUrl:
          type: string
        paymentMethod:
          type: string
          enum: [esewa, khalti, ime-pay]
        status:
          type: string
          enum: [pending, verified, rejected]
        adminNote:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderId, screenshotUrl, paymentMethod, status, createdAt]
    UploadPaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        screenshotUrl:
          type: string
        paymentMethod:
          type: string
          enum: [esewa, khalti, ime-pay]
      required: [orderId, screenshotUrl, paymentMethod]
    VerifyPaymentBody:
      type: object
      properties:
        status:
          type: string
          enum: [verified, rejected]
        adminNote:
          type: string
          nullable: true
      required: [status]
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        expiresAt:
          type: string
          nullable: true
        usageLimit:
          type: integer
          nullable: true
        usageCount:
          type: integer
        active:
          type: boolean
        applicableProductIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
      required: [id, code, discountType, discountValue, usageCount, active, applicableProductIds, createdAt]
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        expiresAt:
          type: string
          nullable: true
        usageLimit:
          type: integer
          nullable: true
        active:
          type: boolean
        applicableProductIds:
          type: array
          items:
            type: integer
      required: [code, discountType, discountValue, active, applicableProductIds]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        productId:
          type: integer
        amount:
          type: number
      required: [code, productId, amount]
    CouponValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        discountAmount:
          type: number
        finalAmount:
          type: number
        message:
          type: string
      required: [valid, discountAmount, finalAmount, message]
...
    PaymentSetting:
      type: object
      properties:
        id:
          type: integer
        method:
          type: string
        label:
          type: string
        enabled:
          type: boolean
        accountName:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        qrImageUrl:
          type: string
          nullable: true
        instructions:
          type: string
          nullable: true
        sortOrder:
          type: integer
        updatedAt:
          type: string
      required: [id, method, label, enabled, sortOrder, updatedAt]
    UpdatePaymentSettingBody:
      type: object
      properties:
        label:
          type: string
        enabled:
          type: boolean
        accountName:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        qrImageUrl:
          type: string
          nullable: true
        instructions:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: []