# lib/db/src/schema/payment-settings.ts
import { pgTable, serial, text, boolean, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const paymentSettingsTable = pgTable("payment_settings", {
  id: serial("id").primaryKey(),
  method: text("method").notNull().unique(),
  label: text("label").notNull(),
  enabled: boolean("enabled").notNull().default(true),
  accountName: text("account_name"),
  accountNumber: text("account_number"),
  qrImageUrl: text("qr_image_url"),
  instructions: text("instructions"),
  sortOrder: integer("sort_order").notNull().default(0),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertPaymentSettingSchema = createInsertSchema(paymentSettingsTable).omit({ id: true, updatedAt: true });
export type InsertPaymentSetting = z.infer<typeof insertPaymentSettingSchema>;
export type PaymentSetting = typeof paymentSettingsTable.$inferSelect;